flask_app/requirements.txt
Flask==3.0.0
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0
